Rename rs_matter::dm::device_types to rs_matter::dm::devices
