+++ examples/src/bin/media_player.rs
@@ -48,7 +48,7 @@ use media_playback::{
 use rs_matter::dm::clusters::desc::{self, ClusterHandler as _};
 use rs_matter::dm::clusters::net_comm::NetworkType;
 use rs_matter::dm::clusters::on_off::{self, ClusterHandler as _, OnOffHandler};
-use rs_matter::dm::device_types::DEV_TYPE_CASTING_VIDEO_PLAYER;
+use rs_matter::dm::devices::DEV_TYPE_CASTING_VIDEO_PLAYER;
 use rs_matter::dm::networks::unix::UnixNetifs;
 use rs_matter::dm::root_endpoint;
 use rs_matter::dm::subscriptions::Subscriptions;
