+++ examples/src/bin/onoff_light.rs
@@ -30,7 +30,7 @@ use log::info;
 use rs_matter::dm::clusters::desc::{self, ClusterHandler as _};
 use rs_matter::dm::clusters::net_comm::NetworkType;
 use rs_matter::dm::clusters::on_off::{self, ClusterHandler as _};
-use rs_matter::dm::device_types::DEV_TYPE_ON_OFF_LIGHT;
+use rs_matter::dm::devices::DEV_TYPE_ON_OFF_LIGHT;
 use rs_matter::dm::networks::unix::UnixNetifs;
 use rs_matter::dm::root_endpoint;
 use rs_matter::dm::subscriptions::Subscriptions;
