+++ examples/src/bin/speaker.rs
@@ -37,7 +37,7 @@ use level_control::{
 use rs_matter::dm::clusters::desc::{self, ClusterHandler as _};
 use rs_matter::dm::clusters::net_comm::NetworkType;
 use rs_matter::dm::clusters::on_off::{ClusterHandler as _, OnOffHandler};
-use rs_matter::dm::device_types::DEV_TYPE_SMART_SPEAKER;
+use rs_matter::dm::devices::DEV_TYPE_SMART_SPEAKER;
 use rs_matter::dm::networks::unix::UnixNetifs;
 use rs_matter::dm::root_endpoint;
 use rs_matter::dm::subscriptions::Subscriptions;
